refactor: decompose container-agent.service.ts into focused sub-services (CQ-007, SL-001)
- Create src/services/container-agent/ directory with 6 sub-services
- SandboxStateManager: owns 3 Maps + 1 Set for state tracking
- WorktreeInitService: worktree creation, path translation
- ContainerExecService: container lifecycle management
- AgentCoreBridgeService: AgentCore start/stop, SSE bridge
- PlanApprovalService: plan ready/approve/reject handlers
- ContainerAgentService facade: ~250 lines, delegates to sub-services
- Replace custom debug/info/warn log with createLogger()
- Backward-compatible re-export maintains all existing imports
